+++ import_test.go
@@ -84,6 +84,18 @@ func Test_fetchModule(t *testing.T) {
 			importPath: "rsc.io/quote",
 			want:       "rsc.io/quote",
 		},
+		{
+			name:       "main module that has nested module inside",
+			root:       "testdata/mod2",
+			importPath: "github.com/hashicorp/vault/api",
+			want:       "github.com/hashicorp/vault/api",
+		},
+		{
+			name:       "module that is nested module inside another main",
+			root:       "testdata/mod2/nestedModule1",
+			importPath: "crawshaw.io/sqlite",
+			want:       "crawshaw.io/sqlite",
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -158,6 +170,21 @@ func Test_getTestModules(t *testing.T) {
 			name: "root with NO ending slash",
 			root: "testdata/mod1",
 			want: []string{"github.com/frankban/quicktest", "github.com/shirou/gopsutil"},
+		},
+		{
+			name: "main module that has nested module inside",
+			root: "testdata/mod2",
+			want: []string{"github.com/shirou/gopsutil", "gopkg.in/natefinch/lumberjack.v2"},
+		},
+		{
+			name: "module that is nested module inside another main",
+			root: "testdata/mod2/nestedModule1",
+			want: []string{"crawshaw.io/sqlite"},
+		},
+		{
+			name: "module that is nested module inside another main take II",
+			root: "testdata/mod2/nestedModule2",
+			want: []string{"github.com/sirupsen/logrus"},
 		},
 	}
 	for _, tt := range tests {