+++ testdata/mod2/nestedModule2/cool/heya/hey.go
@@ -1,0 +1,10 @@
+package heya
+
+import (
+	"github.com/imroc/req"
+)
+
+func Heya() {
+
+	_ = req.LreqBody
+}