+++ testdata/mod2/nestedModule2/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/aerth/spriteutil"
+)
+
+func main() {
+	fmt.Println("mod2/nestedModule2")
+	_ = spriteutil.AnimatedSprite{}
+}