+++ testdata/mod2/somedir/some.go
@@ -1,0 +1,9 @@
+package some
+
+import (
+	"github.com/chromedp/chromedp"
+)
+
+func SomeApi() chromedp.Browser {
+	return chromedp.Browser{}
+}