+++ testdata/mod2/somedir/some_test.go
@@ -1,0 +1,12 @@
+package some
+
+import (
+	"testing"
+
+	"gopkg.in/natefinch/lumberjack.v2"
+)
+
+func TestBaa(t *testing.T) {
+
+	_ = lumberjack.Logger{}
+}